# Speed up rabbit managements tests