--- a/deps/rabbitmq_management/test/clustering_prop_SUITE.erl
+++ b/deps/rabbitmq_management/test/clustering_prop_SUITE.erl
@@ -115,6 +115,9 @@ prop_connection_channel_counts(Config) ->
                 force_stats(Config),
                 Res = validate_counts(Config, Cons),
                 cleanup(Cons),
+                rabbit_ct_helpers:await_condition(
+                  fun () -> validate_counts(Config, []) end,
+                  60000),
                 force_stats(Config),
                 Res
             end).