main: `gmake -C deps/rabbitmq_management tests  16.75s user 12.51s system 3% cpu 13:12.46 total`

this PR: `gmake -C deps/rabbitmq_management tests  15.21s user 10.73s system 8% cpu 4:48.54 total`